migrate from sqlite to mysql

+++ src/db/accountStationOption.ts
@@ -1,6 +1,7 @@
 import { handleDBError } from "../lib/utils";
 import { getDatabase } from "./db";
 import { AccountStationOption, AccountStationOptionInput } from "./dbTypes";
+import { RowDataPacket, ResultSetHeader } from "mysql2/promise";
 
 // Insert a new account station option
 export async function insertAccountStationOptionDataAccess({
@@ -11,18 +12,19 @@ export async function insertAccountStationOptionDataAccess({
   error?: string;
 }> {
   const db = await getDatabase();
+  if (!db) {
+    throw new Error("DB is not defined");
+  }
 
   try {
-    const result = await db.run(
+    const [result] = await db.execute<ResultSetHeader>(
       `INSERT INTO AccountStationOptions (AccountStationOptionName) VALUES (?)`,
       [AccountStationOptionName]
     );
 
-    return { success: true, lastID: result.lastID };
+    return { success: true, lastID: result.insertId };
   } catch (error: unknown) {
     return handleDBError(error);
-  } finally {
-    await db.close();
   }
 }
 
@@ -32,20 +34,21 @@ export async function updateAccountStationOptionDataAccess(
   AccountStationOptionName: string
 ): Promise<{ success: boolean; error?: string }> {
   const db = await getDatabase();
+  if (!db) {
+    throw new Error("DB is not defined");
+  }
 
   try {
-    const result = await db.run(
+    const [result] = await db.execute<ResultSetHeader>(
       `UPDATE AccountStationOptions SET AccountStationOptionName = ? WHERE AccountStationOptionID = ?`,
       [AccountStationOptionName, accountStationOptionID]
     );
 
-    const changes = result.changes ?? 0; // Fallback to 0 if changes is undefined
+    const changes = result.affectedRows ?? 0; // Fallback to 0 if changes is undefined
 
     return { success: changes > 0 };
   } catch (error: unknown) {
     return handleDBError(error);
-  } finally {
-    await db.close();
   }
 }
 
@@ -54,20 +57,21 @@ export async function deleteAccountStationOptionDataAccess(
   accountStationOptionID: number
 ): Promise<{ success: boolean; error?: string }> {
   const db = await getDatabase();
+  if (!db) {
+    throw new Error("DB is not defined");
+  }
 
   try {
-    const result = await db.run(
+    const [result] = await db.execute<ResultSetHeader>(
       `DELETE FROM AccountStationOptions WHERE AccountStationOptionID = ?`,
       [accountStationOptionID]
     );
 
-    const changes = result.changes ?? 0; // Fallback to 0 if changes is undefined
+    const changes = result.affectedRows ?? 0; // Fallback to 0 if changes is undefined
 
     return { success: changes > 0 };
   } catch (error: unknown) {
     return handleDBError(error);
-  } finally {
-    await db.close();
   }
 }
 
@@ -76,14 +80,17 @@ export async function getAllAccountStationOptionsDataAccess(): Promise<
   AccountStationOption[]
 > {
   const db = await getDatabase();
+  if (!db) {
+    throw new Error("DB is not defined");
+  }
 
   try {
-    const rows = await db.all(`SELECT * FROM AccountStationOptions`);
+    const [rows] = await db.execute<RowDataPacket[]>(
+      "SELECT * FROM AccountStationOptions"
+    );
     return rows as AccountStationOption[];
   } catch (error: unknown) {
     return handleDBError(error);
-  } finally {
-    await db.close();
   }
 }
 
@@ -93,16 +100,19 @@ export async function deleteAllAccountStationOptionsDataAccess(): Promise<{
   error?: string;
 }> {
   const db = await getDatabase();
+  if (!db) {
+    throw new Error("DB is not defined");
+  }
 
   try {
-    const result = await db.run(`DELETE FROM AccountStationOptions`);
-    const changes = result.changes ?? 0; // Fallback to 0 if changes is undefined
+    const [result] = await db.execute<ResultSetHeader>(
+      "DELETE FROM AccountStationOptions"
+    );
+    const changes = result.affectedRows ?? 0; // Fallback to 0 if changes is undefined
 
     return { success: changes > 0 };
   } catch (error: unknown) {
     return handleDBError(error);
-  } finally {
-    await db.close();
   }
 }
 
@@ -111,17 +121,18 @@ export async function findAccountStationOptionDataAccess(
   accountStationOptionID: number
 ): Promise<AccountStationOption | null> {
   const db = await getDatabase();
+  if (!db) {
+    throw new Error("DB is not defined");
+  }
 
   try {
-    const row = await db.get(
+    const [rows] = await db.execute<RowDataPacket[]>(
       `SELECT * FROM AccountStationOptions WHERE AccountStationOptionID = ?`,
       [accountStationOptionID]
     );
-    return row ? (row as AccountStationOption) : null;
+    return rows.length > 0 ? (rows[0] as AccountStationOption) : null;
   } catch (error: unknown) {
     return handleDBError(error);
-  } finally {
-    await db.close();
   }
 }
 
@@ -130,16 +141,17 @@ export async function doesAccountStationOptionExistDataAccess(
   accountStationOptionID: number
 ): Promise<boolean> {
   const db = await getDatabase();
+  if (!db) {
+    throw new Error("DB is not defined");
+  }
 
   try {
-    const row = await db.get(
+    const [rows] = await db.execute<RowDataPacket[]>(
       `SELECT 1 FROM AccountStationOptions WHERE AccountStationOptionID = ?`,
       [accountStationOptionID]
     );
-    return !!row; // Return true if the row exists, false otherwise
+    return rows.length > 0; // Return true if the row exists, false otherwise
   } catch (error: unknown) {
     return handleDBError(error);
-  } finally {
-    await db.close();
   }
 }
